Encryption: seprate the cipher method from AESCTRCipherStream to support encrypt/decrypt simple data
